+++ src/test/java/com/hiberus/university/selenium/stepdefs/LoginPageSteps.java
@@ -1,5 +1,6 @@
 package com.hiberus.university.selenium.stepdefs;
 
+import com.hiberus.university.selenium.pages.InventoryPage;
 import com.hiberus.university.selenium.pages.LoginPage;
 import com.hiberus.university.selenium.pages.PagesFactory;
 import io.cucumber.java.en.And;
@@ -7,6 +8,8 @@ import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import lombok.extern.slf4j.Slf4j;
+import org.junit.Assert;
+import org.openqa.selenium.remote.RemoteWebDriver;
 
 @Slf4j
 public class LoginPageSteps {
@@ -20,15 +23,34 @@ public class LoginPageSteps {
   }
 
   @And("the user provides the username {string} and password {string}")
-  public void theUserProvidesTheUsernameAndPassword(String arg0, String arg1) {
+  public void theUserProvidesTheUsernameAndPassword(String username, String password) {
+    PagesFactory pf = PagesFactory.getInstance();
+    log.info("The user provides the username and password");
+
+    LoginPage loginPage = pf.getLoginPage();
+    loginPage.enterUsername(username);
+    loginPage.enterPassword(password);
   }
 
   @When("the user clicks the login button")
   public void theUserClicksTheLoginButton() {
+    PagesFactory pf = PagesFactory.getInstance();
+    log.info("The user clicks the 'Login' button");
+
+    LoginPage loginPage = pf.getLoginPage();
+    loginPage.clickLogin();
   }
 
   @Then("the user is logged successfully and is into the inventory page")
   public void theUserIsLoggedSuccessfullyAndIsIntoTheInventoryPage() {
+    PagesFactory pf = PagesFactory.getInstance();
+    log.info("The user should login successfully and is brought to the inventory page");
+
+    InventoryPage inventoryPage = pf.getInventoryPage();
+    inventoryPage.waitForPageLoad();
+
+    String currentUrl = PagesFactory.getInstance().getDriver().getCurrentUrl();
+    Assert.assertEquals("the URL is not inventory Page", InventoryPage.PAGE_URL, currentUrl);
   }
 
   @Then("The user should be shown an invalid message")
